requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-choices==1.5.0
django==1.11
djangorestframework==3.5.4
flake8==3.3.0
mccabe==0.6.1             # via flake8
mypy==0.501
psycopg2==2.7.1
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
pytz==2017.2              # via django
typed-ast==1.0.3          # via mypy
